Move analysis package into internal
