+++ compiler/internal/analysis/bool.go
