+++ compiler/internal/analysis/escape.go
