+++ compiler/internal/analysis/info.go
