+++ compiler/internal/analysis/info_test.go
