+++ compiler/internal/analysis/sideeffect.go
